worker/jobs/index: Extract `NormalizeIndex` into dedicated module
